Fix ESLint errors: resolve dependency cycle and set-state-in-effect
Extract SpeziContext into its own module to break the circular import
chain (Toaster → SpeziProvider → messages → ... → Toaster). Derive
resolvedTheme synchronously instead of setting state inside useEffect.
Remove duplicate ThemeMode re-export from useTheme.
